+++ src/lightning_app/cli/cmd_install.py
@@ -5,6 +5,7 @@ import subprocess
 import sys
 from typing import Dict, Optional, Tuple
 
+import click
 import requests
 from packaging.version import Version
 
@@ -12,6 +13,83 @@ from lightning_app.core.constants import LIGHTNING_APPS_PUBLIC_REGISTRY, LIGHTNI
 from lightning_app.utilities.app_helpers import Logger
 
 logger = Logger(__name__)
+
+
+@click.group(name="install")
+def install() -> None:
+    """Install Lightning AI selfresources."""
+    pass
+
+
+@install.command("app")
+@click.argument("name", type=str)
+@click.option(
+    "--yes",
+    "-y",
+    is_flag=True,
+    help="disables prompt to ask permission to create env and run install cmds",
+)
+@click.option(
+    "--version",
+    "-v",
+    type=str,
+    help="Specify the version to install. By default it uses 'latest'",
+    default="latest",
+    show_default=True,
+)
+@click.option(
+    "--overwrite",
+    "-f",
+    is_flag=True,
+    default=False,
+    help="When set, overwrite the app directory without asking if it already exists.",
+)
+def install_app(name: str, yes: bool, version: str, overwrite: bool = False) -> None:
+    _install_app_command(name, yes, version, overwrite=overwrite)
+
+
+@install.command("component")
+@click.argument("name", type=str)
+@click.option(
+    "--yes",
+    "-y",
+    is_flag=True,
+    help="disables prompt to ask permission to create env and run install cmds",
+)
+@click.option(
+    "--version",
+    "-v",
+    type=str,
+    help="Specify the version to install. By default it uses 'latest'",
+    default="latest",
+    show_default=True,
+)
+def install_component(name: str, yes: bool, version: str) -> None:
+    _install_component_command(name, yes, version)
+
+
+def _install_app_command(name: str, yes: bool, version: str, overwrite: bool = False) -> None:
+    if "github.com" in name:
+        if version != "latest":
+            logger.warn(
+                f"The provided version {version} isn't the officially supported one. "
+                f"The provided version will be ignored."
+            )
+        return non_gallery_app(name, yes, overwrite=overwrite)
+    else:
+        return gallery_app(name, yes, version, overwrite=overwrite)
+
+
+def _install_component_command(name: str, yes: bool, version: str, overwrite: bool = False) -> None:
+    if "github.com" in name:
+        if version != "latest":
+            logger.warn(
+                f"The provided version {version} isn't the officially supported one. "
+                f"The provided version will be ignored."
+            )
+        return non_gallery_component(name, yes, overwrite=overwrite)
+    else:
+        return gallery_component(name, yes, version, overwrite=overwrite)
 
 
 def gallery_apps_and_components(
@@ -29,7 +107,7 @@ def gallery_apps_and_components(
             entry, app_or_component, org, yes_arg, resource_type="app"
         )
         # run installation if requested
-        _install_app(source_url, git_url, folder_name, cwd=cwd, overwrite=overwrite, git_sha=git_sha)
+        _install_app_from_source(source_url, git_url, folder_name, cwd=cwd, overwrite=overwrite, git_sha=git_sha)
 
         return os.path.join(os.getcwd(), folder_name, entry["appEntrypointFile"])
 
@@ -38,7 +116,7 @@ def gallery_apps_and_components(
         git_url = _show_install_component_prompt(entry, app_or_component, org, yes_arg)
 
         # run installation if requested
-        _install_component(git_url)
+        _install_component_from_source(git_url)
 
         return os.path.join(os.getcwd(), entry["appEntrypointFile"])
 
@@ -60,7 +138,7 @@ def gallery_component(name: str, yes_arg: bool, version_arg: str, cwd: str = Non
     git_url = _show_install_component_prompt(component_entry, component, org, yes_arg)
 
     # run installation if requested
-    _install_component(git_url)
+    _install_component_from_source(git_url)
 
     return os.path.join(os.getcwd(), component_entry["appEntrypointFile"])
 
@@ -71,7 +149,7 @@ def non_gallery_component(gh_url: str, yes_arg: bool, cwd: str = None) -> None:
     git_url = _show_non_gallery_install_component_prompt(gh_url, yes_arg)
 
     # run installation if requested
-    _install_component(git_url)
+    _install_component_from_source(git_url)
 
 
 def gallery_app(name: str, yes_arg: bool, version_arg: str, cwd: str = None, overwrite: bool = False) -> str:
@@ -91,7 +169,7 @@ def gallery_app(name: str, yes_arg: bool, version_arg: str, cwd: str = None, ove
     )
 
     # run installation if requested
-    _install_app(source_url, git_url, folder_name, cwd=cwd, overwrite=overwrite, git_sha=git_sha)
+    _install_app_from_source(source_url, git_url, folder_name, cwd=cwd, overwrite=overwrite, git_sha=git_sha)
 
     return os.path.join(os.getcwd(), folder_name, app_entry["appEntrypointFile"])
 
@@ -102,7 +180,7 @@ def non_gallery_app(gh_url: str, yes_arg: bool, cwd: str = None, overwrite: bool
     repo_url, folder_name = _show_non_gallery_install_app_prompt(gh_url, yes_arg)
 
     # run installation if requested
-    _install_app(repo_url, repo_url, folder_name, cwd=cwd, overwrite=overwrite)
+    _install_app_from_source(repo_url, repo_url, folder_name, cwd=cwd, overwrite=overwrite)
 
 
 def _show_install_component_prompt(entry: Dict[str, str], component: str, org: str, yes_arg: bool) -> str:
@@ -441,7 +519,7 @@ def _install_with_env(repo_url: str, folder_name: str, cwd: str = None) -> None:
     logger.info(m)
 
 
-def _install_app(
+def _install_app_from_source(
     source_url: str, git_url: str, folder_name: str, cwd: str = None, overwrite: bool = False, git_sha: str = None
 ) -> None:
     """Installing lighting app from the `git_url`
@@ -518,7 +596,7 @@ def _install_app(
     logger.info(m)
 
 
-def _install_component(git_url: str) -> None:
+def _install_component_from_source(git_url: str) -> None:
     logger.info("⚡ RUN: pip install")
 
     out = subprocess.check_output(["pip", "install", git_url])
